// ng-app/src/app/types/inventory.type.ts
/**
 * Copyright (c) 2021 Patrik Forsberg <patrik.forsberg@coldmind.com> - All Rights Reserved
 * Unauthorized copying of this file, via any medium is strictly prohibited
 * Proprietary and confidential.
 */

export interface IInventory {
	inventory: IInventoryEntry[];
}

export interface IInventoryEntry {
	art_id: number;
	name:   string;
	stock:  number;
}